# ARROW-13252: [C++] Add offset to CSV error reporting