--- a/cpp/src/arrow/dataset/file_csv.cc
+++ b/cpp/src/arrow/dataset/file_csv.cc
@@ -67,7 +67,7 @@ Result<std::unordered_set<std::string>> GetColumnNames(
   uint32_t parsed_size = 0;
   int32_t max_num_rows = read_options.skip_rows + 1;
   csv::BlockParser parser(pool, parse_options, /*num_cols=*/-1, /*first_row=*/1,
-                          max_num_rows);
+                          /*offset=*/0, max_num_rows);
 
   RETURN_NOT_OK(parser.Parse(util::string_view{first_block}, &parsed_size));