… cpp/src/arrow/csv/parser_test.cc
… python/pyarrow/tests/test_csv.py