@pitrou 2021-10-05
> Most have no significant difference but ParseCSVQuotedBlock, ParseCSVEscapedBlock and ParseCSVFlightsExample did show performance degradation

Sorry for not answering earlier; but the degradation looks quite concerning indeed. A minor difference (e.g. 3-4%) would be ok, but we seem to be getting a 30% slowdown in some of these cases.