@pitrou 2021-11-10
While the micro-benchmarks don't show any regression, it's quite conceivable that a more memory-heavy workload would show a regression. After all, we're building up an additional buffer, even though it's smaller than the values buffer. So I'm a bit lukewarm on this, especially as the motivation is a rather niche use case.