@n3world 2021-11-10
> While the micro-benchmarks don't show any regression, it's quite conceivable that a more memory-heavy workload would show a regression.

That could be an issue if the system is memory constrained relative to the size of the csv

> So I'm a bit lukewarm on this, especially as the motivation is a rather niche use case.

I'm not sure how returning the offset of an error is a niche use case especially when using threaded parsing the line number is not known so when an error occurs there is no line number to indicate where the error was encountered