@pitrou 2021-11-10
> That could be an issue if the system is memory constrained relative to the size of the csv

More generally though, more allocations and more memory writing can increase the impact on the CPU cache and memory subsystem, even if lots of RAM are still available.