@n3world 2022-02-09
It seems like this functionality is not desired. Should I just close the PR and associated jira ticket?