@pitrou 2022-02-09
@n3world Sorry. Judging by the amount of complexity and runtime overhead it adds, I think it's not desirable indeed.